refactor funtion of exec_tree and add comment
